+++ tests/integration/io/iomanager_tests_runner.py
@@ -19,8 +19,14 @@ def check_output(name, file_object, expected):
 
     print("{BLUE}[RUN     ]{END} Checking for \"{expected}\" in {name}".format(name=name, expected=expected, **formatters))
 
-    contents = file_object.readline().rstrip()
-    if (contents != expected):
+
+    found = False
+
+    for line in file_object:
+        if (expected in line.rstrip()):
+            found = True
+
+    if (not found):
         print("{RED}[   ERROR]{END} Got {contents}".format(contents=contents, expected=expected, **formatters))
         errors = errors + 1
     else:
@@ -80,12 +86,12 @@ def run_io_test(test_env, file_uid, expect_logging, expect_replay):
 
     check_output('stderr', error, 'testing error stream')
 
-    output_filename = 'umpire_io_tests.0.{pid}.{uid}.log'.format(uid=file_uid, pid=pid)
-    replay_filename = 'umpire_io_tests.0.{pid}.{uid}.replay'.format(uid=file_uid, pid=pid)
+    output_filename = 'umpire_io_tests.{pid}.{uid}.log'.format(uid=file_uid, pid=pid)
+    replay_filename = 'umpire_io_tests.{pid}.{uid}.replay'.format(uid=file_uid, pid=pid)
 
     if 'UMPIRE_OUTPUT_DIR' in test_env.keys():
-        output_filename = '{dir}/umpire_io_tests.0.{pid}.{uid}.log'.format(dir=test_env['UMPIRE_OUTPUT_DIR'], uid=file_uid, pid=pid)
-        replay_filename = '{dir}/umpire_io_tests.0.{pid}.{uid}.replay'.format(dir=test_env['UMPIRE_OUTPUT_DIR'], uid=file_uid, pid=pid)
+        output_filename = '{dir}/umpire_io_tests.{pid}.{uid}.log'.format(dir=test_env['UMPIRE_OUTPUT_DIR'], uid=file_uid, pid=pid)
+        replay_filename = '{dir}/umpire_io_tests.{pid}.{uid}.replay'.format(dir=test_env['UMPIRE_OUTPUT_DIR'], uid=file_uid, pid=pid)
 
     if expect_logging:
         check_file_exists(output_filename)